rename geocoding helper to geocoder and namespace it
